Move the availability-gate tests into their own module
`tests/test_capabilities.py` sits against the 1 MB `check-added-large-files` cap (#7304).
It passed on this branch in isolation but failed on the PR merge ref, where main's own
additions to the same file push the merged result over: 1,048,551 (branch) + 1,848 (main
since the merge base) = 1,050,399 bytes.

Moves `TestUnavailableCapabilityToolsAreNotCallable`, the reveal-rejection test, and the
`secret_op` helpers they share -- availability is one coherent subject, so it earns a
module rather than an arbitrary slice. `TestPrepareToolsHook` stays put: it sits between
them in the file but tests the hook, not the gate.

867 test ids before, 867 after, none renamed. Leaves ~10 KB of headroom on the merged
file, so #7304 (a real split of what remains) still stands.
