+++ tests/test_capabilities.py
@@ -78,7 +78,6 @@ from pydantic_ai.exceptions import (
 from pydantic_ai.messages import (
     AgentStreamEvent,
     BinaryImage,
-    CompactionPart,
     EnqueuedMessagesEvent,
     FilePart,
     FunctionToolCallEvent,
@@ -7144,50 +7143,6 @@ class TestSkipModelRequestInteraction:
         assert output == 'model short-circuited'
 
 
-def secret_op() -> str:
-    """A capability-owned tool an owner's `prepare_tools` filters out."""
-    return 'SECRET EXECUTED'  # pragma: no cover
-
-
-def _report_secret_op_outcome(messages: list[ModelMessage]) -> ModelResponse | None:
-    """Report the blocked call, once the retry prompt refusing it is in history.
-
-    `secret_op` never executes in these tests — that's the point — so a tool return for it would
-    itself be the failure, and is left to `secret_op`'s own unreachable body to catch.
-    """
-    for part in iter_message_parts(messages, ModelRequest, RetryPromptPart):
-        return make_text_response(f'BLOCKED: {part.content}')
-    return None
-
-
-def _call_secret_op(messages: list[ModelMessage], info: AgentInfo) -> ModelResponse:
-    """Directly call `secret_op` without it ever being offered."""
-    return _report_secret_op_outcome(messages) or ModelResponse(
-        parts=[ToolCallPart(tool_name='secret_op', args={}, tool_call_id='s1')]
-    )
-
-
-def _load_compact_then_call_secret_op(messages: list[ModelMessage], info: AgentInfo) -> ModelResponse:
-    """Load `guarded`, emit a `CompactionPart` to reset that state, then direct-call its tool."""
-    if (report := _report_secret_op_outcome(messages)) is not None:
-        return report
-
-    parts = [part for message in messages for part in message.parts]
-    if not any(isinstance(p, ToolReturnPart) and p.tool_name == LOAD_CAPABILITY_TOOL_NAME for p in parts):
-        return ModelResponse(
-            parts=[ToolCallPart(tool_name=LOAD_CAPABILITY_TOOL_NAME, args={'id': 'guarded'}, tool_call_id='l1')]
-        )
-    if not any(isinstance(p, CompactionPart) for p in parts):
-        # Compaction lands in its own step, alongside a harmless call so the run continues.
-        return ModelResponse(
-            parts=[
-                CompactionPart(content='Summary: guarded was loaded earlier.', provider_name='function'),
-                ToolCallPart(tool_name='ping', args={}, tool_call_id='p1'),
-            ]
-        )
-    return ModelResponse(parts=[ToolCallPart(tool_name='secret_op', args={}, tool_call_id='s1')])
-
-
 class TestPrepareToolsHook:
     async def test_filter_function_tools(self):
         """Capability can filter out function tools by name."""
@@ -7301,209 +7256,6 @@ class TestPrepareToolsHook:
         result = await agent.run('hello')
         # A runs first, then B, so suffix order is _A_B
         assert 'desc_A_B' in result.output
-
-
-class TestUnavailableCapabilityToolsAreNotCallable:
-    """A deferred capability's tools cannot be called until the capability is available.
-
-    Availability is the gate: the capability's instructions and hooks arrive as a bundle when it
-    loads, so calling one of its tools earlier would run it without the context the model was meant
-    to have read first. The refusal is a retry the model can act on, not an "unknown tool" dead end.
-    """
-
-    @staticmethod
-    def _guarded_capability() -> Capability[Any]:
-        return Capability[Any](
-            id='guarded',
-            description='Guarded tools.',
-            toolsets=[FunctionToolset([secret_op])],
-            defer_loading=True,
-        )
-
-    async def test_never_loaded_capability_tool_is_refused(self):
-        """A direct call to an unloaded capability's tool is refused, and says how to fix it."""
-        agent = Agent(FunctionModel(_call_secret_op), capabilities=[self._guarded_capability()])
-
-        result = await agent.run('hello')
-
-        assert result.output == snapshot(
-            "BLOCKED: Tool 'secret_op' is not available yet: it belongs to capability 'guarded'. "
-            'Call `load_capability` for it first, then call the tool on a later turn, once the '
-            "capability's instructions are in view."
-        )
-
-    async def test_capability_tool_is_refused_again_after_compaction(self):
-        """A `CompactionPart` resets the load state, so the tool needs loading again."""
-        agent = Agent(FunctionModel(_load_compact_then_call_secret_op), capabilities=[self._guarded_capability()])
-
-        @agent.tool_plain
-        def ping() -> str:
-            return 'pong'
-
-        result = await agent.run('hello')
-
-        assert 'is not available yet' in result.output
-
-    async def test_available_capability_does_not_excuse_an_undiscovered_tool(self):
-        """An always-on capability's search-gated tool still has to be searched for.
-
-        The capability being available makes its tools eligible to be shown, not evidence that any
-        of them were — so availability must not short-circuit the discovery requirement.
-        """
-        toolset = FunctionToolset[Any]()
-        toolset.add_function(secret_op, name='secret_op', defer_loading=True)
-        eager = Capability[Any](id='eager', description='Always on.', toolsets=[toolset], defer_loading=False)
-
-        agent = Agent(FunctionModel(_call_secret_op), capabilities=[eager])
-
-        result = await agent.run('hello')
-
-        assert result.output == snapshot(
-            "BLOCKED: Tool 'secret_op' is not available yet: search for it first, then call it once "
-            'the search result has shown you its schema.'
-        )
-
-    async def test_an_availability_refusal_leaves_room_to_act_on_it(self):
-        """The refusal grants one attempt beyond the tool's budget, and no more.
-
-        It is not a mistake about the call's arguments — it says the run isn't in a state where the
-        tool can be called, and names the step that fixes it. On the default budget of 1, charging
-        it like a validation error would make a single act of disobedience fatal, so a message
-        written to be acted on would never get the chance. The ceiling stays finite: a model that
-        never takes the correction still ends the run.
-        """
-
-        def call_twice_then_report(messages: list[ModelMessage], info: AgentInfo) -> ModelResponse:
-            refusals = list(iter_message_parts(messages, ModelRequest, RetryPromptPart))
-            if len(refusals) >= 2:
-                return make_text_response(f'refused {len(refusals)}x, still running')
-            return ModelResponse(parts=[ToolCallPart(tool_name='secret_op', args={}, tool_call_id=f's{len(refusals)}')])
-
-        agent = Agent(FunctionModel(call_twice_then_report), capabilities=[self._guarded_capability()])
-        assert (await agent.run('hello')).output == snapshot('refused 2x, still running')
-
-        def keep_calling(messages: list[ModelMessage], info: AgentInfo) -> ModelResponse:
-            refusals = list(iter_message_parts(messages, ModelRequest, RetryPromptPart))
-            return ModelResponse(parts=[ToolCallPart(tool_name='secret_op', args={}, tool_call_id=f's{len(refusals)}')])
-
-        stubborn = Agent(FunctionModel(keep_calling), capabilities=[self._guarded_capability()])
-        with pytest.raises(UnexpectedModelBehavior, match="Tool 'secret_op' exceeded max retries"):
-            await stubborn.run('hello')
-
-    async def test_an_availability_refusal_does_not_spend_the_budget_a_real_failure_needs(self):
-        """A refusal must not leave the tool with nothing left when it is later called properly.
-
-        The refusal says the run isn't in a state where the tool can be called. Charging it to
-        `retries` would mean a tool refused once, then loaded and called correctly, aborts on its
-        *first* genuine failure with no retry at all — the budget spent on a state problem rather
-        than on the mistake it exists for.
-        """
-        kinds: list[str] = []
-
-        def load_then_fail(messages: list[ModelMessage], info: AgentInfo) -> ModelResponse:
-            parts = [part for message in messages for part in message.parts]
-            loaded = any(
-                isinstance(part, ToolReturnPart) and part.tool_name == LOAD_CAPABILITY_TOOL_NAME for part in parts
-            )
-            kinds[:] = [
-                'availability' if 'is not available yet' in str(part.content) else 'real'
-                for part in iter_message_parts(messages, ModelRequest, RetryPromptPart)
-            ]
-            if not loaded and not kinds:
-                # Call before loading: refused for availability.
-                return ModelResponse(parts=[ToolCallPart(tool_name='failing_op', args={}, tool_call_id='early')])
-            if not loaded:
-                return ModelResponse(
-                    parts=[ToolCallPart(tool_name=LOAD_CAPABILITY_TOOL_NAME, args={'id': 'guarded'}, tool_call_id='l')]
-                )
-            if 'real' in kinds:
-                return make_text_response('the real retry survived')
-            return ModelResponse(parts=[ToolCallPart(tool_name='failing_op', args={}, tool_call_id='real')])
-
-        def failing_op() -> str:
-            raise ModelRetry('give me better arguments')
-
-        toolset = FunctionToolset[Any]([failing_op])
-        capability = Capability[Any](id='guarded', description='Guarded.', toolsets=[toolset], defer_loading=True)
-
-        agent = Agent(FunctionModel(load_then_fail), capabilities=[capability])
-        result = await agent.run('hello')
-
-        assert result.output == snapshot('the real retry survived')
-        # The availability refusal came first and cost nothing; the genuine failure still got its retry.
-        assert kinds == snapshot(['availability', 'real'])
-
-    async def test_an_availability_refusal_is_charged_against_the_tools_own_budget(self):
-        """A tool's configured `max_retries` governs its refusals, not the manager's default.
-
-        The refusal is raised while resolving, before the caller has bound the resolved tool, so
-        the budget could easily be taken from the manager default that an unresolvable name gets.
-        A tool that asked for a larger budget must still get it.
-        """
-        refusals = 0
-
-        def keep_calling(messages: list[ModelMessage], info: AgentInfo) -> ModelResponse:
-            nonlocal refusals
-            refusals = len(list(iter_message_parts(messages, ModelRequest, RetryPromptPart)))
-            return ModelResponse(parts=[ToolCallPart(tool_name='secret_op', args={}, tool_call_id=f's{refusals}')])
-
-        toolset = FunctionToolset[Any]()
-        toolset.add_function(secret_op, name='secret_op', retries=4)
-        capability = Capability[Any](id='guarded', description='Guarded.', toolsets=[toolset], defer_loading=True)
-
-        agent = Agent(FunctionModel(keep_calling), capabilities=[capability])
-        with pytest.raises(UnexpectedModelBehavior, match='exceeded max retries count of 4'):
-            await agent.run('hello')
-
-        # The tool's own budget of 4, plus the one extra an availability refusal is granted.
-        assert refusals == snapshot(5)
-
-    async def test_unavailable_capability_tool_is_not_advertised(self):
-        """The tool is withheld as well as uncallable — never visible-but-unusable."""
-        advertised: list[list[str]] = []
-
-        def model_fn(messages: list[ModelMessage], info: AgentInfo) -> ModelResponse:
-            advertised.append(sorted(t.name for t in info.function_tools))
-            return make_text_response('done')
-
-        agent = Agent(FunctionModel(model_fn), capabilities=[self._guarded_capability()])
-
-        @agent.tool_plain
-        def untouched() -> str:
-            return 'safe'  # pragma: no cover
-
-        await agent.run('hello')
-
-        assert advertised == snapshot([['load_capability', 'untouched']])
-
-
-async def test_reveal_of_another_capabilitys_tool_is_rejected_even_while_loaded():
-    """Being loaded exempts a capability's *own* tools, never another capability's.
-
-    `load_capability` is allowed past the guard for the bundle it is activating, so the guard has to
-    stay strict about everything else — otherwise one loaded capability would become a route to
-    revealing any other capability's tools without activating them.
-    """
-
-    def other_op() -> str:
-        return 'OTHER'  # pragma: no cover
-
-    other = Capability[Any](
-        id='other', description='Other tools.', toolsets=[FunctionToolset([other_op])], defer_loading=True
-    )
-
-    def smuggler() -> ToolReturn:
-        return ToolReturn(return_value='ok', tools=['other_op'])
-
-    smuggling = Capability[Any](id='smuggling', description='Smuggling tools.', toolsets=[FunctionToolset([smuggler])])
-
-    def model_fn(messages: list[ModelMessage], info: AgentInfo) -> ModelResponse:
-        return ModelResponse(parts=[ToolCallPart(tool_name='smuggler', args={}, tool_call_id='s1')])
-
-    agent = Agent(FunctionModel(model_fn), capabilities=[other, smuggling])
-
-    with pytest.raises(UserError, match="cannot reveal 'other_op'"):
-        await agent.run('go')
 
 
 class TestPrepareOutputToolsHook:
